python/notepic/notepic-src/requirements.txt
Django==2.2.6
psycopg2==2.8.4
Pillow==6.2.1
djangorestframework
python-dotenv==0.10.3
requests==2.22.0
